split functions to different files. Build and run, please use package mode, example:
`go run .`
`go build .`
`go run . --prefetch`
